performed new api request with favourite, adjusted way viewmodel saves data, layout adjustments, etc.
