+++ app/src/main/res/layout/activity_sun.xml
@@ -90,7 +90,7 @@
 
             <androidx.recyclerview.widget.RecyclerView
                 android:id="@+id/sunsetFavsRecyclerView"
-                android:layout_width="409dp"
+                android:layout_width="0dp"
                 android:layout_height="0dp"
                 android:layout_marginTop="32dp"
                 app:layout_constraintBottom_toBottomOf="parent"
