+++ app/src/main/res/values-fr-rCA/strings_sunset.xml
@@ -36,7 +36,7 @@
     <string name="sunset_golden_hour_value">L\'heure d\'or va ici</string>
     <string name="sunset_day_length_value">La longueur du jour va ici</string>
     <string name="sunset_fav_title">Favori</string>
-    <string name="sunset_help_main">Saisissez une latitude et une longitude et effectuez une recherche pour voir les données relatives au coucher et au lever du soleil dans un fragment. Appuyez sur le bouton \"favori\" pour ajouter un élément à la liste des favoris, et appuyez sur \"supprimer\" pour retirer l\'élément des favoris.</string>
+    <string name="sunset_help_main">Saisissez une latitude et une longitude et effectuez une recherche pour connaître les détails du coucher et du lever du soleil. Appuyez sur le bouton favori dans l\'écran des détails pour enregistrer un emplacement favori. Utilisez le bouton Supprimer dans la barre d\'outils pour supprimer un élément favori.</string>
     <string name="sunset_ok">OK</string>
     <string name="sunset_del_fav_msg">Voulez-vous supprimer les favoris #</string>
     <string name="sunset_del_fav_title">Supprimer le favori</string>
@@ -53,4 +53,5 @@
     <string name="sunset_lng_detail_value">Lng</string>
     <string name="sunset_details_fav_title">Emplacement préféré</string>
     <string name="sunsnet_close_details">Fermer</string>
+    <string name="sunset_no_search_terms">Veuillez saisir une latitude et une longitude</string>
 </resources>
