+++ app/src/main/res/values/colors.xml
@@ -2,4 +2,5 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+    <color name="sunset_highlight">#FFD0B3</color>
 </resources>
